requirements.txt
mcp>=1.0.0
httpx>=0.27.0
urllib3>=2.0.0
pyyaml>=6.0.0
